Feat: 티끌 데이터 UI 바인딩

+++ Tikkeul/Presentation/Sceenes/Record/Model/CategoryRecordData.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct CategoryRecordData {
     let category: TikkeulCategory
-    let value: Double
+    let value: Int
     let money: Int
 }
